requirements.txt
python-dotenv
markdown
requests